more cleanup on bard.  increase volume (x3), fix setting of partial minute values in tune gui

+++ src/main/java/net/shadowmage/ancientwarfare/npc/gui/GuiNpcBard.java
@@ -141,7 +141,7 @@ private int addTuneEntry(final BardTuneEntry entry, final int index, int startHe
     @Override
     public void onValueUpdated(float value)
       {
-      entry.setLength((int)value);
+      entry.setLength(value);
       }
     };
   area.addGuiElement(length);
